+++ packages/hono-takibi/src/cli/index.ts
@@ -1,1 +1,40 @@
-export { honoTakibi } from './hono-takibi/index.js'
+import { asyncAndThen, ok, type Result } from '../result/index.js'
+import { isHelpRequested, sliceArgv } from '../utils/index.js'
+
+import { parseCli } from './parse.js'
+import { takibi } from './takibi.js'
+
+const HELP_TEXT = `Usage: hono-takibi <input.{yaml,json,tsp}> -o <routes.ts> [options]
+
+Options:
+  --export-type        export TypeScript type aliases
+  --export-schema      export Zod schema objects
+  --template           generate app file and handler stubs
+  --test               generate empty *.test.ts files
+  --base-path <path>   api prefix (default: /)
+  -h, --help           display help for command`
+
+/**
+ * CLI entry point for `hono-takibi`.
+ *
+ * @returns A `Result` containing help text or CLI execution result.
+ */
+export async function honoTakibi(): Promise<Result<{ message: string }, string>> {
+  const args = sliceArgv(process.argv)
+  return isHelpRequested(sliceArgv(process.argv))
+    ? ok({ message: HELP_TEXT })
+    : await asyncAndThen(parseCli(args), async (cli) =>
+        asyncAndThen(
+          await takibi(
+            cli.input,
+            cli.output,
+            cli.exportSchema ?? false,
+            cli.exportType ?? false,
+            cli.template ?? false,
+            cli.test ?? false,
+            cli.basePath,
+          ),
+          async (result) => ok(result),
+        ),
+      )
+}
