--- packages/hono-takibi/src/cli/parse/index.ts
@@ -1,2 +1,0 @@
-export { parseCli } from './parse-cli.js'
-export { parseIO } from './parse-io.js'
